Add the currently-interpreted method to the top of the stack trace.
This is currently the best this stack trace implementation can get, because of the execution architecture. Imagine that an interpreted method calls another interpreted method.
What you have is an interpreter context that does an ObjC message send, that sets up a new FFI trampoline, that creates a new interpreter context and interprets the inner method.
There's no way from debugging the inner method to see what outer methods called it.
I think that the solution is to introduce the Interpreter abstraction, and then have a per-thread stack of interpreters. Then the debugger could match up calls to -executeInContext: with interpreters in the stack.